Extracted json_utils.hpp from utils.hpp
