+++ src/cpp/src/llm_pipeline_static.cpp
@@ -13,6 +13,7 @@
 #include <jinja2cpp/user_callable.h>
 
 #include "text_callback_streamer.hpp"
+#include "json_utils.hpp"
 #include "utils.hpp"
 
 namespace {
