+++ src/cpp/src/text2image/schedulers/euler_discrete.cpp
@@ -9,7 +9,7 @@
 #include <random>
 
 #include "text2image/numpy_utils.hpp"
-#include "utils.hpp"
+#include "json_utils.hpp"
 
 namespace ov {
 namespace genai {
