+++ src/cpp/src/text2image/schedulers/lms_discrete.cpp
@@ -6,8 +6,7 @@
 #include <cmath>
 #include <fstream>
 
-#include "utils.hpp"
-
+#include "json_utils.hpp"
 namespace {
 
 // https://gist.github.com/lorenzoriano/5414671
