+++ src/cpp/src/text2image/stable_diffusion_pipeline.hpp
@@ -6,7 +6,7 @@
 #include <ctime>
 #include <cassert>
 
-#include "utils.hpp"
+#include "json_utils.hpp"
 #include "lora_helper.hpp"
 
 namespace ov {
