+++ src/cpp/src/text2image/stable_diffusion_xl_pipeline.hpp
@@ -6,7 +6,7 @@
 #include <ctime>
 #include <cassert>
 
-#include "utils.hpp"
+#include "json_utils.hpp"
 
 namespace ov {
 namespace genai {
