+++ src/cpp/src/utils.hpp
@@ -3,8 +3,6 @@
 
 #pragma once
 
-#include <nlohmann/json.hpp>
-
 #include "openvino/genai/llm_pipeline.hpp"
 #include "visual_language/processor_config.hpp"
 
@@ -23,36 +21,6 @@ void initialize_position_ids(ov::Tensor& position_ids, const ov::Tensor& attenti
 ov::Tensor extend_attention(ov::Tensor attention_mask);
 
 void update_position_ids(ov::Tensor&& position_ids, const ov::Tensor&& attention_mask);
-
-/// @brief reads value to param if T argument type is aligned with value stores in json
-/// if types are not compatible leave param unchanged
-template <typename T>
-void read_json_param(const nlohmann::json& data, const std::string& name, T& param) {
-    if (data.contains(name)) {
-        if (data[name].is_number() || data[name].is_boolean() || data[name].is_string() || data[name].is_object()) {
-            param = data[name].get<T>();
-        }
-    } else if (name.find(".") != std::string::npos) {
-        size_t delimiter_pos = name.find(".");
-        std::string key = name.substr(0, delimiter_pos);
-        if (!data.contains(key)) {
-            return;
-        }
-        std::string rest_key = name.substr(delimiter_pos + 1);
-
-        read_json_param(data[key], rest_key, param);
-    }
-}
-
-template <typename V>
-void read_json_param(const nlohmann::json& data, const std::string& name, std::vector<V>& param) {
-    if (data.contains(name) && data[name].is_array()) {
-        param.resize(0);
-        for (const auto elem : data[name]) {
-            param.push_back(elem.get<V>());
-        }
-    }
-}
 
 template <typename T>
 void read_anymap_param(const ov::AnyMap& config_map, const std::string& name, T& param) {
