+++ src/cpp/src/visual_language/processor_config.cpp
@@ -2,7 +2,8 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #include "processor_config.hpp"
-#include "utils.hpp"
+#include "json_utils.hpp"
+
 #include <fstream>
 
 ov::genai::ProcessorConfig::ProcessorConfig(const std::filesystem::path& json_path) {
