+++ src/cpp/src/whisper_generation_config.cpp
@@ -9,6 +9,7 @@
 #include <openvino/runtime/core.hpp>
 
 #include "utils.hpp"
+#include "json_utils.hpp"
 
 namespace ov {
 namespace genai {
